mwinda-user/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
psycopg2-binary
email-validator>=2.0
pydantic_settings
asyncpg

passlib
bcrypt
python-jose[cryptography]
aio_pika
psycopg[binary]
python-barcode
bcrypt==4.0.1
passlib[bcrypt]==1.7.4